Follows more traditional service worker lifecycle.
